requirements.txt
asgiref==3.2.7
Django==3.0.4
django-filter==2.2.0
Pillow==7.1.1
pytz==2019.3
Reqtxt==1.0
sqlparse==0.3.1
